refactor: markdown codeblockwrapper now exported from multi-code-block.js
