+++ packages/ui-kit/src/components/markdown.js
@@ -3,7 +3,7 @@ import { css } from '@emotion/core';
 import styled from '@emotion/styled';
 import RibbonIcon from '../icons/ribbon-icon.svg';
 import { colors, dimensions, typography, tokens } from '../design-system';
-import { CodeBlockMarkdownWrapper as CodeBlock } from './code-block';
+import { CodeBlockMarkdownWrapper as CodeBlock } from './multi-code-block';
 
 const TypographyPage = styled.div`
   font-family: ${typography.fontFamilies.primary};
